strategy fix for resolving path and assumptions
