+++ src/metldata/builtin_transformations/common/utils.py
@@ -22,7 +22,16 @@ from typing import Any
 
 from schemapack import dumps_datapack, dumps_schemapack
 from schemapack.spec.datapack import DataPack
-from schemapack.spec.schemapack import ClassDefinition, SchemaPack
+from schemapack.spec.schemapack import (
+    ClassDefinition,
+    Relation,
+    SchemaPack,
+)
+
+from metldata.builtin_transformations.common.path.path_elements import (
+    RelationPathElement,
+    RelationPathElementType,
+)
 
 
 def model_to_dict(model: SchemaPack) -> dict[str, Any]:
@@ -39,3 +48,18 @@ def content_to_dict(class_def: ClassDefinition) -> dict[str, Any]:
     """Converts a content schema into a dictionary."""
     class_def_dict = json.loads(class_def.model_dump_json())
     return class_def_dict["content"]
+
+
+def get_relation(element: RelationPathElement, schema: SchemaPack) -> Relation:
+    """Get the relation object for a path element.
+
+    Args:
+        element: The path element to get the relation for.
+        schema: The underlying schema.
+
+    Returns:
+        The relation object.
+    """
+    element_active = element.type_ == RelationPathElementType.ACTIVE
+    class_name = element.source if element_active else element.target
+    return schema.classes[class_name].relations[element.property]
