+++ src/metldata/builtin_transformations/count_references/assumptions.py
@@ -20,9 +20,8 @@ from schemapack.spec.schemapack import SchemaPack
 from metldata.builtin_transformations.common.assumptions import (
     assert_class_is_source,
     assert_object_path_exists,
-    assert_only_direct_relations,
     assert_path_classes_and_relations_exist,
-    assert_target_multiplicity,
+    assert_relation_target_multiplicity,
 )
 from metldata.builtin_transformations.count_references.instruction import (
     AddReferenceCountPropertyInstruction,
@@ -38,7 +37,6 @@ def check_model_assumptions(
         for instruction in instructions:
             path = instruction.source_relation_path
             assert_path_classes_and_relations_exist(model=schema, path=path)
-            assert_only_direct_relations(path=path)
             assert_class_is_source(path=path, instruction=instruction)
             assert_object_path_exists(model=schema, instruction=instruction)
-            assert_target_multiplicity(model=schema, path=path)
+            assert_relation_target_multiplicity(model=schema, path=path)
