+++ tests/fixtures/example_transformations/count_content_values/multiple/config.yaml
@@ -4,7 +4,7 @@ count_content_values:
       object_path: samples_summary.stats
       property_name: sex
     source:
-      relation_path: Dataset(individuals)>Individual
+      relation_path: Dataset(biospecimens)>Biospecimen(individuals)>Individual
       content_path: sex
   - class_name: Dataset
     target_content:
