+++ tests/fixtures/example_transformations/count_content_values/multiple/transformed.datapack.yaml
@@ -13,9 +13,6 @@ resources:
               blood: 2 # <-
               subcutaneous_adipose_tissue: 1 # <-
       relations:
-        individuals:
-          - indv_a
-          - indv_b
         biospecimens:
           - bios_a
           - bios_b
