+++ tests/fixtures/example_transformations/count_content_values/multiple/transformed.schemapack.yaml
@@ -44,14 +44,6 @@ classes:
         "type": "object",
       }
     relations:
-      individuals:
-        targetClass: Individual
-        multiple:
-          origin: true
-          target: true
-        mandatory:
-          origin: false
-          target: true
       biospecimens:
         targetClass: Biospecimen
         multiple:
