+++ tests/fixtures/example_transformations/count_references/multiple/transformed.datapack.yaml
@@ -7,24 +7,30 @@ resources:
         format: FASTQ
         checksum: 1a5ac10ab42911dc0224172c118a326d9a4c03969112a2f3eb1ad971e96e92b8
         size: 12321
+        dataset_summary:
+          count: 1 # <-
     file_b:
       content:
         filename: file_b.fastq
         format: FASTQ
         checksum: 2b5ac10ab42911dc0224172c118a326d9a4c03969112a2f3eb1ad971e96e92c9
         size: 12314
+        dataset_summary:
+          count: 1 # <-
     file_c:
       content:
         filename: file_c.fastq
         format: FASTQ
         checksum: a9c24870071da03f78515e6197048f3a2172e90e597e9250cd01a0cb8f0986ed
         size: 12123
+        dataset_summary:
+          count: 1 # <-
   Dataset:
     dataset_1:
       content:
         dac_contact: dac@example.org
-        file_summary: # <-
-          count: 3
+        sample_summary: # <-
+          count: 2
       relations:
         files:
           - file_a
@@ -34,22 +40,19 @@ resources:
     sample_x:
       content:
         description: Some sample.
-        file_summary: # <-
-          count: 2
       relations:
         files:
           - file_a
           - file_b
     sample_y:
-      content:
-        file_summary: # <-
-          count: 1
+      content: {}
       relations:
         files:
           - file_c
   Experiment:
     experiment_i:
       content:
+        description: Some experiment.
         sample_summary: # <-
           count: 2
       relations:
