+++ tests/fixtures/example_transformations/count_references/multiple/transformed.schemapack.yaml
@@ -4,7 +4,28 @@ classes:
   File:
     id:
       propertyName: alias
-    content: ../../../example_content_schemas/File.schema.json
+    content:
+      {
+        "$schema": "http://json-schema.org/draft-07/schema#",
+        "additionalProperties": false,
+        "description": "A file is an object that contains information generated from a process, either an Experiment or an Analysis.",
+        "properties":
+          {
+            "checksum": { "type": "string" },
+            "filename": { "type": "string" },
+            "format": { "type": "string" },
+            "size": { "type": "integer" },
+            "dataset_summary":
+              {
+                "type": "object",
+                "additionalProperties": true,
+                "properties": { "count": { "type": "integer" } },
+              },
+          },
+        "required":
+          ["filename", "format", "checksum", "size", "dataset_summary"],
+        "type": "object",
+      }
   Dataset:
     id:
       propertyName: alias
@@ -16,14 +37,14 @@ classes:
         "properties":
           {
             "dac_contact": { "type": "string" },
-            "file_summary":
+            "sample_summary":
               {
                 "type": "object",
                 "additionalProperties": true,
                 "properties": { "count": { "type": "integer" } },
               },
           },
-        "required": ["file_summary"],
+        "required": ["sample_summary"],
         "type": "object",
       }
     relations:
@@ -38,24 +59,7 @@ classes:
   Sample:
     id:
       propertyName: alias
-    content:
-      {
-        "$schema": "http://json-schema.org/draft-07/schema#",
-        "additionalProperties": false,
-        "description": "A sample used to generate files in the context of an experiment.",
-        "properties":
-          {
-            "description": { "type": "string" },
-            "file_summary":
-              {
-                "type": "object",
-                "additionalProperties": true,
-                "properties": { "count": { "type": "integer" } },
-              },
-          },
-        "required": ["file_summary"],
-        "type": "object",
-      }
+    content: ../../../example_content_schemas/Sample.schema.json
     relations:
       files:
         targetClass: File
